Added Header, Footer, CreatePoll, and Home Structure
Created components for the header and footer and put them in the index file so they will be above and below every page that loads.

Created Home module and added routing so it will be loaded in between headers and footers.

Created CreatePoll module and set up routing to it as well as a test routerLink to it from home.